Regorganised code
