+++ src/main/java/seedu/saveit/logic/commands/report/PrintReportCommand.java
@@ -8,6 +8,19 @@ import static seedu.saveit.logic.parser.CliSyntax.PREFIX_START_DATE;
 
 import java.util.HashMap;
 
+import javafx.print.PageLayout;
+import javafx.print.PrinterJob;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.SnapshotParameters;
+import javafx.scene.chart.BarChart;
+import javafx.scene.chart.Chart;
+import javafx.scene.chart.PieChart;
+import javafx.scene.image.ImageView;
+import javafx.scene.image.WritableImage;
+import javafx.scene.transform.Scale;
+
 import seedu.saveit.logic.commands.Command;
 import seedu.saveit.logic.commands.CommandResult;
 import seedu.saveit.logic.commands.exceptions.CommandException;
@@ -17,6 +30,7 @@ import seedu.saveit.model.report.Report;
 import seedu.saveit.ui.Bar;
 import seedu.saveit.ui.Graph;
 import seedu.saveit.ui.Pie;
+import seedu.saveit.ui.exceptions.PrinterException;
 
 
 /**
@@ -53,6 +67,69 @@ public class PrintReportCommand extends Command {
         this.toPrint = toPrint;
     }
 
+    /**
+     * Invokes printer job of Javafx.
+     *
+     * @throws PrinterException if job cannot finish.
+     */
+    public void printerJob(Graph graph) throws CommandException {
+
+        try {
+            PrinterJob printerJob = PrinterJob.createPrinterJob();
+            PageLayout pageLayout = printerJob.getJobSettings().getPageLayout();
+
+            WritableImage snapshot = snapshot(graph);
+
+            ImageView ivSnapshot = new ImageView(snapshot);
+            double scaleX = pageLayout.getPrintableWidth() / ivSnapshot.getImage().getWidth();
+            double scaleY = pageLayout.getPrintableHeight() / ivSnapshot.getImage().getHeight();
+            double scale = Math.min(scaleX, scaleY);
+            if (scale < 1.0) {
+                ivSnapshot.getTransforms().add(new Scale(scale, scale));
+            }
+            boolean jobStatus = printerJob.printPage(ivSnapshot);
+            if (jobStatus) {
+                printerJob.endJob();
+            } else {
+                printerJob.cancelJob();
+                throw new PrinterException("Set available printer as default "
+                        + "printer before printing");
+            }
+
+        } catch (Exception e) {
+            throw new CommandException("Set available printer as default "
+                    + "printer before printing.");
+        }
+    }
+
+
+    /**
+     * Takes a snapshot of the graph
+     * to be exported.
+     *
+     * @return image of the graph.
+     */
+    public WritableImage snapshot(Graph graph) {
+        Node node = (Node) graph.constructGraph();
+        Scene sc = new Scene((Parent) node, 800, 600);
+        Chart chart = null;
+
+        if (node instanceof PieChart) {
+            chart = (PieChart) node;
+
+        } else if (node instanceof BarChart) {
+            chart = (BarChart) node;
+        }
+
+        assert chart != null;
+
+        chart.setAnimated(false);
+        WritableImage img = new WritableImage(800, 600);
+        node.snapshot(new SnapshotParameters(), img);
+
+        return img;
+    }
+
     @Override
     public CommandResult execute(Model model) throws CommandException {
 
@@ -73,6 +150,8 @@ public class PrintReportCommand extends Command {
         } else {
             throw new CommandException(MESSAGE_FAIL);
         }
+
+        printerJob(graph);
 
         return new CommandResult(MESSAGE_SUCCESS, graph, false, false, true);
     }
