+++ src/main/java/seedu/saveit/logic/commands/report/ReportWindowHelpCommand.java
@@ -16,6 +16,6 @@ public class ReportWindowHelpCommand extends ReportCommand {
 
     @Override
     public ReportCommandResult execute(Model model) {
-        return new ReportCommandResult(SHOWING_HELP_MESSAGE, true);
+        return new ReportCommandResult(SHOWING_HELP_MESSAGE);
     }
 }
