+++ src/test/java/seedu/saveit/logic/commands/report/ExportReportCommandTest.java
@@ -13,21 +13,15 @@ import static seedu.saveit.logic.commands.CommandTestUtil.VALID_ORGANISATION_TAG
 import static seedu.saveit.logic.commands.CommandTestUtil.VALID_START_DATE_BUS;
 import static seedu.saveit.logic.commands.CommandTestUtil.VALID_START_DATE_BUS_ALT;
 import static seedu.saveit.logic.commands.CommandTestUtil.assertCommandFailure;
-import static seedu.saveit.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.saveit.testutil.Assert.assertThrows;
 import static seedu.saveit.testutil.TypicalAccounts.getTypicalAccountList;
 
-import java.util.HashMap;
-
 import org.junit.jupiter.api.Test;
 
-import seedu.saveit.logic.commands.CommandResult;
 import seedu.saveit.model.Model;
 import seedu.saveit.model.ModelManager;
 import seedu.saveit.model.UserPrefs;
-import seedu.saveit.testutil.ExportFileBuilder;
 import seedu.saveit.testutil.ReportBuilder;
-import seedu.saveit.ui.Pie;
 
 public class ExportReportCommandTest {
 
@@ -79,28 +73,6 @@ public class ExportReportCommandTest {
 
     }
 
-    @Test
-    public void executeSuccess() {
-
-        ReportBuilder rp = new ReportBuilder()
-                .withStartDate(VALID_START_DATE_BUS)
-                .withEndDate(VALID_END_DATE_BUS)
-                .withGraphType(VALID_GRAPH_BAR_CAPS)
-                .withOrganise(VALID_ORGANISATION_MONTH);
-
-        ExportFileBuilder f = new ExportFileBuilder()
-                .withFileName(VALID_FILE_NAME)
-                .withGraph(new Pie(new HashMap(), VALID_ORGANISATION_MONTH));
-
-
-        Model model = new ModelManager(getTypicalAccountList(), new UserPrefs());
-
-        ExportReportCommand command1 = new ExportReportCommand(rp.build(), VALID_FILE_NAME);
-        CommandResult r = new CommandResult(ExportReportCommand.MESSAGE_SUCCESS, f.build(), true);
-
-        assertCommandSuccess(command1, model, r, model);
-
-    }
 
     @Test
     public void executeFailure() {
